[노은수] refactor : 질문 요청에 대한 훅을 최초 요청과 무한 스크롤 요청으로 분리
